pyproject.toml
[project]
name = "blizzapi"
version = "1.0.0"
authors = [
  { name="Nick Stuer", email="nickstuer@duck.com" },
]
description = "Python package to connect to the blizzard API."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
]
dependencies = [
  "keyring>=25.6.0",
  "requests-oauthlib>=2.0.0",
]

[tool.pytest.ini_options]
pythonpath = [
  "src/"
]
testpaths = [
    "tests/",
]

[project.urls]
Homepage = "https://github.com/nickstuer/blizzapi"
Issues = "https://github.com/nickstuer/blizzapi/issues"

[tool.ruff]
line-length = 120
show-fixes = true

[tool.ruff.lint]
select = [
    "ALL",
]
ignore = [
  "D100",
  "D101",
  "D102",
  "D103",
  "D107",
  "G004",
  "ANN204",
  "INP001",
  "D104",
  "D213",
  "D203",
  "COM812",
  "N803"
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101", "D103", "PLR2004", "ANN201", "ANN202", "ANN001", "D401", "D404"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"


[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.2.1",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: not covered",
    "@overload",
    "if TYPE_CHECKING:",
    "def __del__",
]

[tool.coverage.run]
branch = true
relative_files = true
omit = [
    ".tox/*",
    "tests/*",
]